+++ app/src/main/java/com/concordium/wallet/ui/cis2/SearchTokenBottomSheet.kt
@@ -61,18 +61,18 @@ class SearchTokenBottomSheet : BaseBottomSheetDialogFragment() {
             }
         })
 
-        binding.search.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
-            override fun onQueryTextSubmit(query: String?): Boolean {
-                tokensAccountDetailsAdapter.dataSet = _viewModel.tokens.value!!.filter { it.token.uppercase().contains(query?.uppercase() ?: "") }.toTypedArray()
-                tokensAccountDetailsAdapter.notifyDataSetChanged()
-                return false
-            }
-            override fun onQueryTextChange(newText: String?): Boolean {
-                tokensAccountDetailsAdapter.dataSet = _viewModel.tokens.value!!.filter { it.token.uppercase().contains(newText?.uppercase() ?: "") }.toTypedArray()
-                tokensAccountDetailsAdapter.notifyDataSetChanged()
-                return false
-            }
-        })
+//        binding.search.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
+//            override fun onQueryTextSubmit(query: String?): Boolean {
+//                tokensAccountDetailsAdapter.dataSet = _viewModel.tokens.value!!.filter { it.token.uppercase().contains(query?.uppercase() ?: "") }.toTypedArray()
+//                tokensAccountDetailsAdapter.notifyDataSetChanged()
+//                return false
+//            }
+//            override fun onQueryTextChange(newText: String?): Boolean {
+//                tokensAccountDetailsAdapter.dataSet = _viewModel.tokens.value!!.filter { it.token.uppercase().contains(newText?.uppercase() ?: "") }.toTypedArray()
+//                tokensAccountDetailsAdapter.notifyDataSetChanged()
+//                return false
+//            }
+//        })
     }
 
     private fun initObservers() {